requirements.txt
typer[all]==0.9.0
rich==13.7.0
PyQt6==6.6.0
httpx==0.25.0
beautifulsoup4==4.12.0
Pillow==10.1.0
reportlab==4.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.0
